pyproject.toml
[tool.poetry]
name = "ia-calendar-api"
version = "0.1.0"
description = ""
authors = ["Franco Anfossi <franco.anfossi@uc.cl>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
databases = "^0.9.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.5.2"
requests = "^2.32.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.6.9"
alembic = "^1.13.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I"]
exclude = [".venv", "migrations", "tests"]
fix = true 
